+++ src/main/java/com/scubakay/autorelog/commands/AutoRelogCommand.java
@@ -1,15 +1,8 @@
 package com.scubakay.autorelog.commands;
 
 import com.mojang.brigadier.CommandDispatcher;
-import com.mojang.brigadier.arguments.IntegerArgumentType;
 import com.mojang.brigadier.context.CommandContext;
-import com.mojang.brigadier.tree.ArgumentCommandNode;
 import com.mojang.brigadier.tree.LiteralCommandNode;
-import com.scubakay.autorelog.AutoRelogClient;
-import com.scubakay.autorelog.commands.suggestions.AfkDelaySuggestionsProvider;
-import com.scubakay.autorelog.commands.suggestions.DelaySuggestionProvider;
-import com.scubakay.autorelog.commands.suggestions.IntervalSuggestionProvider;
-import com.scubakay.autorelog.commands.suggestions.MaxAttemptsSuggestionProvider;
 import com.scubakay.autorelog.util.Reconnect;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager;
 import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
@@ -23,102 +16,7 @@ public class AutoRelogCommand {
                 .executes(AutoRelogCommand::autorelog)
                 .build();
 
-        LiteralCommandNode<FabricClientCommandSource> delayNode = ClientCommandManager
-                .literal("delay")
-                .build();
-
-        LiteralCommandNode<FabricClientCommandSource> configNode = ClientCommandManager
-                .literal("config")
-                .build();
-
-        ArgumentCommandNode<FabricClientCommandSource, Integer> delayArgumentNode = ClientCommandManager
-                .argument("delay", IntegerArgumentType.integer())
-                .suggests(new DelaySuggestionProvider())
-                .executes(ctx -> delay(ctx, IntegerArgumentType.getInteger(ctx, "delay")))
-                .build();
-
-        LiteralCommandNode<FabricClientCommandSource> intervalNode = ClientCommandManager
-                .literal("interval")
-                .build();
-
-        ArgumentCommandNode<FabricClientCommandSource, Integer> intervalArgumentNode = ClientCommandManager
-                .argument("interval", IntegerArgumentType.integer())
-                .suggests(new IntervalSuggestionProvider())
-                .executes(ctx -> interval(ctx, IntegerArgumentType.getInteger(ctx, "interval")))
-                .build();
-
-        LiteralCommandNode<FabricClientCommandSource> maxAttemptsNode = ClientCommandManager
-                .literal("maxAttempts")
-                .build();
-
-        ArgumentCommandNode<FabricClientCommandSource, Integer> maxAttemptsArgumentNode = ClientCommandManager
-                .argument("maxAttempts", IntegerArgumentType.integer())
-                .suggests(new MaxAttemptsSuggestionProvider())
-                .executes(ctx -> maxAttempts(ctx, IntegerArgumentType.getInteger(ctx, "maxAttempts")))
-                .build();
-
-        LiteralCommandNode<FabricClientCommandSource> cancelNode = ClientCommandManager
-                .literal("cancel")
-                .executes(AutoRelogCommand::cancel)
-                .build();
-
-        LiteralCommandNode<FabricClientCommandSource> modeNode = ClientCommandManager
-                .literal("mode")
-                .build();
-
-        LiteralCommandNode<FabricClientCommandSource> manualModeNode = ClientCommandManager
-                .literal("manual")
-                .executes(AutoRelogCommand::manualMode)
-                .build();
-
-        LiteralCommandNode<FabricClientCommandSource> automaticModeNode = ClientCommandManager
-                .literal("automatic")
-                .executes(AutoRelogCommand::afkDetectionMode)
-                .build();
-
-        LiteralCommandNode<FabricClientCommandSource> afkDelayNode = ClientCommandManager
-                .literal("afkDelay")
-                .build();
-
-        ArgumentCommandNode<FabricClientCommandSource, Integer> afkDelayArgumentNode = ClientCommandManager
-                .argument("afkDelay", IntegerArgumentType.integer())
-                .suggests(new AfkDelaySuggestionsProvider())
-                .executes(ctx -> afkDelay(ctx, IntegerArgumentType.getInteger(ctx, "afkDelay")))
-                .build();
-
-        LiteralCommandNode<FabricClientCommandSource> loggingNode = ClientCommandManager
-                .literal("logging")
-                .executes(AutoRelogCommand::logging)
-                .build();
-
         dispatcher.getRoot().addChild(autoRelogNode);
-
-        // Add config node
-        autoRelogNode.addChild(configNode);
-
-        // Add delay config node
-        configNode.addChild(delayNode);
-        delayNode.addChild(delayArgumentNode);
-
-        // Add delay config node
-        configNode.addChild(intervalNode);
-        intervalNode.addChild(intervalArgumentNode);
-
-        // Add logging config node
-        configNode.addChild(loggingNode);
-
-        // Add max attempts config node
-        configNode.addChild(maxAttemptsNode);
-        maxAttemptsNode.addChild(maxAttemptsArgumentNode);
-
-        configNode.addChild(modeNode);
-        modeNode.addChild(manualModeNode);
-        modeNode.addChild(automaticModeNode);
-
-        configNode.addChild(afkDelayNode);
-        afkDelayNode.addChild(afkDelayArgumentNode);
-
-        autoRelogNode.addChild(cancelNode);
     }
 
     private static int autorelog(CommandContext<FabricClientCommandSource> context) {
@@ -126,77 +24,13 @@ public class AutoRelogCommand {
             context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_error_singleplayer"), false);
             return -1;
         }
-        Reconnect.getInstance().activate();
-        context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_activated"), false);
-        return 1;
-    }
-
-    private static int cancel(CommandContext<FabricClientCommandSource> context) {
-        if (context.getSource().getClient().isInSingleplayer()) {
-            context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_error_singleplayer"), false);
-            return -1;
+        if (Reconnect.getInstance().isActive()) {
+            Reconnect.getInstance().deactivate();
+            context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_deactivated"), false);
+        } else {
+            Reconnect.getInstance().activate();
+            context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_activated"), false);
         }
-        Reconnect.getInstance().deactivate();
-        context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_deactivated"), false);
-        return 1;
-    }
-
-    private static int delay(CommandContext<FabricClientCommandSource> context, int delay) {
-        if (delay < 1) {
-            context.getSource().getPlayer().sendMessage(Text.translatable("commands.config_error_delay"), false);
-            return -1;
-        }
-        AutoRelogClient.CONFIG.setDelay(delay);
-        context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_delay_changed", delay), false);
-        return 1;
-    }
-
-    private static int interval(CommandContext<FabricClientCommandSource> context, int interval) {
-        if (interval < 1) {
-            context.getSource().getPlayer().sendMessage(Text.translatable("commands.config_error_interval"), false);
-            return -1;
-        }
-        AutoRelogClient.CONFIG.setInterval(interval);
-        context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_interval_changed", interval), false);
-        return 1;
-    }
-
-    private static int maxAttempts(CommandContext<FabricClientCommandSource> context, int maxAttempts) {
-        if (maxAttempts < 0) {
-            context.getSource().getPlayer().sendMessage(Text.translatable("commands.config_error_max_attempts"), false);
-            return -1;
-        }
-        AutoRelogClient.CONFIG.setMaxAttempts(maxAttempts);
-        context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_max_attempts_changed", maxAttempts), false);
-        return 1;
-    }
-
-    private static int logging(CommandContext<FabricClientCommandSource> context) {
-        AutoRelogClient.CONFIG.setLogging(!AutoRelogClient.CONFIG.isLogging());
-        context.getSource().getPlayer().sendMessage(Text.translatable("commands.logging.toggled", AutoRelogClient.CONFIG.isLogging() ? "Enabled" : "Disabled"), false);
-        return 1;
-    }
-
-    private static int afkDelay(CommandContext<FabricClientCommandSource> context, int afkDelay) {
-        if (afkDelay < 0) {
-            context.getSource().getPlayer().sendMessage(Text.translatable("commands.config_error_delay"), false);
-            return -1;
-        }
-        AutoRelogClient.CONFIG.setAfkDelay(afkDelay);
-        context.getSource().getPlayer().sendMessage(Text.translatable("commands.afkDetection.delayChanged", afkDelay), false);
-        return 1;
-    }
-
-    private static int manualMode(CommandContext<FabricClientCommandSource> context) {
-        if(AutoRelogClient.CONFIG.isAfkDetection()) Reconnect.getInstance().deactivate();
-        AutoRelogClient.CONFIG.setAfkDetection(false);
-        context.getSource().getPlayer().sendMessage(Text.translatable("commands.mode.manual"), false);
-        return 1;
-    }
-
-    private static int afkDetectionMode(CommandContext<FabricClientCommandSource> context) {
-        AutoRelogClient.CONFIG.setAfkDetection(true);
-        context.getSource().getPlayer().sendMessage(Text.translatable("commands.mode.afkDetection"), false);
         return 1;
     }
 }