+++ src/main/java/com/scubakay/autorelog/util/AfkMode.java
@@ -1,0 +1,6 @@
+package com.scubakay.autorelog.util;
+
+public enum AfkMode {
+    AUTOMATIC,
+    MANUAL
+}